fetch: do this project

+++ src/main/java/ru/practicum/shareit/user/UserServiceImpl.java
@@ -3,7 +3,7 @@ package ru.practicum.shareit.user;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import ru.practicum.shareit.exception.ConflictException;
+import ru.practicum.shareit.exception.NotFoundException;
 import ru.practicum.shareit.item.ItemStorage;
 import ru.practicum.shareit.exception.BadRequestException;
 import ru.practicum.shareit.user.dto.UserDto;
@@ -16,56 +16,47 @@ import java.util.List;
 @RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
     @Autowired
-    private final UserStorage userStorageImpl;
+    private final UserRepository userRepository;
     @Autowired
     private final ItemStorage inMemoryItemStorage;
 
-    private long id = 0;
-
     @Override
     public User create(UserDto userDto) {
-        if (userStorageImpl.checkIsEmailExist(userDto)) {
-            throw UtilityStuff.logError(new ConflictException("Пользователь с адресом электронной почты " +
-                    userDto.getEmail() + " уже существует."));
-        }
-        User user = UserMapper.fromUserDto(userDto);
-        id++;
-        user.setId(id);
-        return userStorageImpl.add(user);
+        return userRepository.save(UserMapper.fromUserDto(userDto));
     }
 
     @Override
     public User getById(long userId) {
-        return userStorageImpl.findById(userId).orElseThrow(() ->
-                UtilityStuff.logError(new BadRequestException("User с id " + userId + " не найден.")));
+        return userRepository.findById(userId).orElseThrow(() ->
+                UtilityStuff.logError(new NotFoundException("User с id " + userId + " не найден.")));
     }
 
     @Override
     public List<User> getAll() {
-        return userStorageImpl.findAll();
+        return userRepository.findAll();
     }
 
     @Override
     public User update(long userId, UserDto userDto) {
-        User user = userStorageImpl.findById(userId).orElseThrow(() ->
-                UtilityStuff.logError(new BadRequestException("User с id " + userId + " не найден.")));
+        User user = userRepository.findById(userId).orElseThrow(() ->
+                UtilityStuff.logError(new NotFoundException("User с id " + userId + " не найден.")));
         if (userDto.getName() != null) {
             user.setName(userDto.getName());
         }
         if (userDto.getEmail() != null) {
-            if (userStorageImpl.checkIsEmailExist(userDto) &&
-                    !userDto.getEmail().equals(userStorageImpl.findById(userId).orElseThrow().getEmail())) {
-                throw UtilityStuff.logError(new ConflictException("Пользователь с адресом электронной почты " +
-                        userDto.getEmail() + " уже существует."));
-            }
             user.setEmail(userDto.getEmail());
         }
-        return userStorageImpl.edit(user);
+        return userRepository.save(user);
     }
 
     @Override
     public void deleteById(long userId) {
-        userStorageImpl.deleteById(userId);
-        inMemoryItemStorage.deleteAllByUserId(userId);
+        if (userRepository.existsById(userId)) {
+            userRepository.deleteById(userId);
+            inMemoryItemStorage.deleteAllByUserId(userId);
+        } else {
+            throw UtilityStuff.logError(new NotFoundException("User с id " + userId + " не найден."));
+        }
+
     }
 }
